requirements.txt
# -------------------------------
# Core framework (Python 3.12)
# -------------------------------
Django==4.2.20
# For openapi/swagger view:
coreapi == 2.3.3

# -------------------------------
# Django REST Framework
# Used by sagenkarta_rest_api and sagendatabas_es_api
# -------------------------------
djangorestframework==3.15.2

# -------------------------------
# OpenAPI / Swagger UI (modern replacement)
# - Replaces deprecated: coreapi, django-rest-swagger
# -------------------------------
#coreapi == 2.3.3
drf-spectacular>=0.27,<0.28
drf-spectacular-sidecar>=2024.3,<2025.0  # ships Swagger UI assets

# If you must keep the old stack, comment the two lines above and
# (not recommended) un-comment these—but they are not supported on Django 4.2:
# coreapi==2.3.3
# django-rest-swagger==2.2.0

# -------------------------------
# Database
# -------------------------------
# mysqlclient: 2.2.x supports Python 3.12; older 2.1.x may fail to build.
#mysqlclient>=2.2,<3.0
mysqlclient==2.2.7

# -------------------------------
# Image handling (required by models.ImageField)
# -------------------------------
#python 3.6:
#Pillow==8.2.0
#Pillow==8.4.0 gives:
#ImageField TradarkAdmin.Persons.image: (fields.E210) Cannot use ImageField because Pillow is not installed.:
# https://pillow.readthedocs.io/en/latest/installation.html
# libjpeg: https://installati.one/ubuntu/22.04/libjpeg-dev/
#python 3.8-10:
#Pillow==9.4.0
#python 3.12:
Pillow==11.3.0
# If you remove ImageField usages (e.g., in sagenkarta_rest_api or TradarkAdmin)
# For django models.ImageField: (sagenkarta_rest_api.Persons.image: (fields.E210) Cannot use ImageField because Pillow is not installed)
# you may also remove Pillow.

# -------------------------------
# HTTP clients / CORS
# -------------------------------
requests==2.32.3
# requests>=2.32 works with urllib3 2.x
urllib3==2.3.0
# CORS
#django-cors-headers==3.10.1
#django-cors-headers==4.7.0
#django-cors-headers>=4.4,<5.0
django-cors-headers==4.9.0

# -------------------------------
# Security headers (optional)
# -------------------------------
# django-csp can cause "Denied by X-Frame-Options: SAMEORIGIN" if misconfigured.
# Uncomment only if you need CSP and configure allowed frame ancestors.
# django-csp==3.7

# -------------------------------
# Reverse proxy (legacy)
# -------------------------------
# django-revproxy is outdated and may require source changes for modern requests/urllib3.
# If you depend on it, you have two realistic paths:

# (A) Keep modern requests/urllib3 and vendor/fix revproxy in your repo (recommended):
# -e git+https://github.com/jazzband/django-revproxy@master#egg=django-revproxy

# (B) Pin older HTTP stack just for revproxy (NOT recommended globally):
# django-revproxy==0.12.0
# requests==2.26.0
# urllib3==1.26.18

# Used in external proxy services as LantmaterietProxyView:
# django-revproxy==0.9.13
# Since django-revproxy 0.13.0 is outdated and not actively maintained, you need to modify the source code of the package.
# OR Downgrade requests to 2.26.0
# There's no django-revproxy-ng package on PyPI.
#django-revproxy-ng==
#django-revproxy==0.12.0
# -e git+https://git@github.com/jazzband/django-revproxy@master#egg=django-revproxy


# -------------------------------
# Notes / migration crumbs
# -------------------------------
# DRF router.register(...):
#   base_name was removed; use 'basename' instead to avoid:
#   TypeError: register() got an unexpected keyword argument 'base_name'

